baml-language: update ty/field attrs for SAP
