+++ baml_language/crates/baml_type/src/attr.rs
@@ -1,6 +1,6 @@
 //! SAP types re-exported from `baml_base`.
 //!
-//! The canonical definitions live in `baml_base::sap` so that both
+//! The canonical definitions live in `baml_base::attr` so that both
 //! `baml_type` and `baml_compiler_tir` can use them without a cyclic dependency.
 
 pub use baml_base::{FieldAttr, SapAttrValue, SapConstValue, TyAttr};
